Complete scanForFaces().
Completes the scanForFaces() method; right now, this was modeled off of the face_detecy.py script that was used in the initial development. I tweaked the parameters of the front_cascade that was used, and it is now much better at only pulling faces from a scene. The class has two other cascades (an eye cascade and alternate face cascade), but they are currently unused, because they did not enhance the performance, or allow filtering of other images that had detected faces.