Complete captureScreen().
Completes captureScreen() function, though it may need a bit of redesign to allow it to 'query' a video stream, as opposed to controlling the entire thing. Completes some light testing of the class: run with 'python TV_Watcher.py'; shows how running a video and capturing scenes, and scanning for images works.